Moved `launcher.py` into the main `psij` directory. Fixes #131.
