+++ src/psij/launchers/__init__.py
@@ -2,10 +2,9 @@
 
 Launcher implementations must be in this package in order for them to be registered automatically.
 Alternatively, you may register a launcher implementation using
-:func:`psij.launchers.launcher.Launcher.register_launcher`.
+:func:`psij.launcher.Launcher.register_launcher`.
 """
 
-from .launcher import Launcher
 from .single import SingleLauncher
 from .multiple import MultipleLauncher
 from .mpirun import MPILauncher
@@ -14,7 +13,6 @@ from .jsrun import JsrunLauncher
 from .aprun import AprunLauncher
 
 __all__ = [
-    "Launcher",
     "SingleLauncher",
     "MultipleLauncher",
     "MPILauncher",
